Move ObjectSpace to test only
