+++ spec/spec_helper.rb
@@ -26,6 +26,7 @@ require 'support/http_helpers'
 require 'support/log_helpers'
 require 'support/metric_helpers'
 require 'support/network_helpers'
+require 'support/object_space_helper'
 require 'support/platform_helpers'
 require 'support/span_helpers'
 require 'support/spy_transport'
